+++ modules/AWSPowerShell/Cmdlets/Connect/Basic/Get-CONNPromptFile-Cmdlet.cs
@@ -22,46 +22,46 @@ using System.Management.Automation;
 using System.Text;
 using Amazon.PowerShell.Common;
 using Amazon.Runtime;
-using Amazon.CloudTrail;
-using Amazon.CloudTrail.Model;
+using Amazon.Connect;
+using Amazon.Connect.Model;
 
-namespace Amazon.PowerShell.Cmdlets.CT
+namespace Amazon.PowerShell.Cmdlets.CONN
 {
     /// <summary>
-    /// Describes the settings for the event selectors that you configured for your trail.
-    /// The information returned for your event selectors includes the following:
-    /// 
-    ///  <ul><li><para>
-    /// If your event selector includes read-only events, write-only events, or all events.
-    /// This applies to both management events and data events.
-    /// </para></li><li><para>
-    /// If your event selector includes management events.
-    /// </para></li><li><para>
-    /// If your event selector includes data events, the resources on which you are logging
-    /// data events.
-    /// </para></li></ul><para>
-    /// For more information about logging management and data events, see the following topics
-    /// in the <i>CloudTrail User Guide</i>:
-    /// </para><ul><li><para><a href="https://docs.aws.amazon.com/awscloudtrail/latest/userguide/logging-management-events-with-cloudtrail.html">Logging
-    /// management events</a></para></li><li><para><a href="https://docs.aws.amazon.com/awscloudtrail/latest/userguide/logging-data-events-with-cloudtrail.html">Logging
-    /// data events</a></para></li></ul>
+    /// Gets the prompt file.
     /// </summary>
-    [Cmdlet("Get", "CTEventSelector")]
-    [OutputType("Amazon.CloudTrail.Model.GetEventSelectorsResponse")]
-    [AWSCmdlet("Calls the AWS CloudTrail GetEventSelectors API operation.", Operation = new[] {"GetEventSelectors"}, SelectReturnType = typeof(Amazon.CloudTrail.Model.GetEventSelectorsResponse), LegacyAlias="Get-CTEventSelectors")]
-    [AWSCmdletOutput("Amazon.CloudTrail.Model.GetEventSelectorsResponse",
-        "This cmdlet returns an Amazon.CloudTrail.Model.GetEventSelectorsResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    [Cmdlet("Get", "CONNPromptFile")]
+    [OutputType("System.String")]
+    [AWSCmdlet("Calls the Amazon Connect Service GetPromptFile API operation.", Operation = new[] {"GetPromptFile"}, SelectReturnType = typeof(Amazon.Connect.Model.GetPromptFileResponse))]
+    [AWSCmdletOutput("System.String or Amazon.Connect.Model.GetPromptFileResponse",
+        "This cmdlet returns a System.String object.",
+        "The service call response (type Amazon.Connect.Model.GetPromptFileResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
     )]
-    public partial class GetCTEventSelectorCmdlet : AmazonCloudTrailClientCmdlet, IExecutor
+    public partial class GetCONNPromptFileCmdlet : AmazonConnectClientCmdlet, IExecutor
     {
         
-        #region Parameter TrailName
+        #region Parameter InstanceId
         /// <summary>
         /// <para>
-        /// <para>Specifies the name of the trail or trail ARN. If you specify a trail name, the string
-        /// must meet the following requirements:</para><ul><li><para>Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_),
-        /// or dashes (-)</para></li><li><para>Start with a letter or number, and end with a letter or number</para></li><li><para>Be between 3 and 128 characters</para></li><li><para>Have no adjacent periods, underscores or dashes. Names like <code>my-_namespace</code>
-        /// and <code>my--namespace</code> are not valid.</para></li><li><para>Not be in IP address format (for example, 192.168.5.4)</para></li></ul><para>If you specify a trail ARN, it must be in the format:</para><para><code>arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail</code></para>
+        /// <para>The identifier of the Amazon Connect instance. You can <a href="https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html">find
+        /// the instance ID</a> in the Amazon Resource Name (ARN) of the instance.</para>
+        /// </para>
+        /// </summary>
+        #if !MODULAR
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        #else
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
+        public System.String InstanceId { get; set; }
+        #endregion
+        
+        #region Parameter PromptId
+        /// <summary>
+        /// <para>
+        /// <para>A unique identifier for the prompt.</para>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -72,26 +72,26 @@ namespace Amazon.PowerShell.Cmdlets.CT
         [System.Management.Automation.AllowNull]
         #endif
         [Amazon.PowerShell.Common.AWSRequiredParameter]
-        public System.String TrailName { get; set; }
+        public System.String PromptId { get; set; }
         #endregion
         
         #region Parameter Select
         /// <summary>
-        /// Use the -Select parameter to control the cmdlet output. The default value is '*'.
-        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.CloudTrail.Model.GetEventSelectorsResponse).
-        /// Specifying the name of a property of type Amazon.CloudTrail.Model.GetEventSelectorsResponse will result in that property being returned.
+        /// Use the -Select parameter to control the cmdlet output. The default value is 'PromptPresignedUrl'.
+        /// Specifying -Select '*' will result in the cmdlet returning the whole service response (Amazon.Connect.Model.GetPromptFileResponse).
+        /// Specifying the name of a property of type Amazon.Connect.Model.GetPromptFileResponse will result in that property being returned.
         /// Specifying -Select '^ParameterName' will result in the cmdlet returning the selected cmdlet parameter value.
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
-        public string Select { get; set; } = "*";
+        public string Select { get; set; } = "PromptPresignedUrl";
         #endregion
         
         #region Parameter PassThru
         /// <summary>
-        /// Changes the cmdlet behavior to return the value passed to the TrailName parameter.
-        /// The -PassThru parameter is deprecated, use -Select '^TrailName' instead. This parameter will be removed in a future version.
+        /// Changes the cmdlet behavior to return the value passed to the PromptId parameter.
+        /// The -PassThru parameter is deprecated, use -Select '^PromptId' instead. This parameter will be removed in a future version.
         /// </summary>
-        [System.Obsolete("The -PassThru parameter is deprecated, use -Select '^TrailName' instead. This parameter will be removed in a future version.")]
+        [System.Obsolete("The -PassThru parameter is deprecated, use -Select '^PromptId' instead. This parameter will be removed in a future version.")]
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
         public SwitchParameter PassThru { get; set; }
         #endregion
@@ -109,7 +109,7 @@ namespace Amazon.PowerShell.Cmdlets.CT
             #pragma warning disable CS0618, CS0612 //A class member was marked with the Obsolete attribute
             if (ParameterWasBound(nameof(this.Select)))
             {
-                context.Select = CreateSelectDelegate<Amazon.CloudTrail.Model.GetEventSelectorsResponse, GetCTEventSelectorCmdlet>(Select) ??
+                context.Select = CreateSelectDelegate<Amazon.Connect.Model.GetPromptFileResponse, GetCONNPromptFileCmdlet>(Select) ??
                     throw new System.ArgumentException("Invalid value for -Select parameter.", nameof(this.Select));
                 if (this.PassThru.IsPresent)
                 {
@@ -118,14 +118,21 @@ namespace Amazon.PowerShell.Cmdlets.CT
             }
             else if (this.PassThru.IsPresent)
             {
-                context.Select = (response, cmdlet) => this.TrailName;
+                context.Select = (response, cmdlet) => this.PromptId;
             }
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
-            context.TrailName = this.TrailName;
+            context.InstanceId = this.InstanceId;
             #if MODULAR
-            if (this.TrailName == null && ParameterWasBound(nameof(this.TrailName)))
+            if (this.InstanceId == null && ParameterWasBound(nameof(this.InstanceId)))
             {
-                WriteWarning("You are passing $null as a value for parameter TrailName which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+                WriteWarning("You are passing $null as a value for parameter InstanceId which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
+            context.PromptId = this.PromptId;
+            #if MODULAR
+            if (this.PromptId == null && ParameterWasBound(nameof(this.PromptId)))
+            {
+                WriteWarning("You are passing $null as a value for parameter PromptId which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
             
@@ -142,11 +149,15 @@ namespace Amazon.PowerShell.Cmdlets.CT
         {
             var cmdletContext = context as CmdletContext;
             // create request
-            var request = new Amazon.CloudTrail.Model.GetEventSelectorsRequest();
+            var request = new Amazon.Connect.Model.GetPromptFileRequest();
             
-            if (cmdletContext.TrailName != null)
+            if (cmdletContext.InstanceId != null)
             {
-                request.TrailName = cmdletContext.TrailName;
+                request.InstanceId = cmdletContext.InstanceId;
+            }
+            if (cmdletContext.PromptId != null)
+            {
+                request.PromptId = cmdletContext.PromptId;
             }
             
             CmdletOutput output;
@@ -181,15 +192,15 @@ namespace Amazon.PowerShell.Cmdlets.CT
         
         #region AWS Service Operation Call
         
-        private Amazon.CloudTrail.Model.GetEventSelectorsResponse CallAWSServiceOperation(IAmazonCloudTrail client, Amazon.CloudTrail.Model.GetEventSelectorsRequest request)
+        private Amazon.Connect.Model.GetPromptFileResponse CallAWSServiceOperation(IAmazonConnect client, Amazon.Connect.Model.GetPromptFileRequest request)
         {
-            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS CloudTrail", "GetEventSelectors");
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "Amazon Connect Service", "GetPromptFile");
             try
             {
                 #if DESKTOP
-                return client.GetEventSelectors(request);
+                return client.GetPromptFile(request);
                 #elif CORECLR
-                return client.GetEventSelectorsAsync(request).GetAwaiter().GetResult();
+                return client.GetPromptFileAsync(request).GetAwaiter().GetResult();
                 #else
                         #error "Unknown build edition"
                 #endif
@@ -209,9 +220,10 @@ namespace Amazon.PowerShell.Cmdlets.CT
         
         internal partial class CmdletContext : ExecutorContext
         {
-            public System.String TrailName { get; set; }
-            public System.Func<Amazon.CloudTrail.Model.GetEventSelectorsResponse, GetCTEventSelectorCmdlet, object> Select { get; set; } =
-                (response, cmdlet) => response;
+            public System.String InstanceId { get; set; }
+            public System.String PromptId { get; set; }
+            public System.Func<Amazon.Connect.Model.GetPromptFileResponse, GetCONNPromptFileCmdlet, object> Select { get; set; } =
+                (response, cmdlet) => response.PromptPresignedUrl;
         }
         
     }