+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2ImportImageTask-Cmdlet.cs
@@ -41,6 +41,8 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     public partial class GetEC2ImportImageTaskCmdlet : AmazonEC2ClientCmdlet, IExecutor
     {
         
+        protected override bool IsSensitiveResponse { get; set; } = true;
+        
         #region Parameter Filter
         /// <summary>
         /// <para>