+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Import-EC2Snapshot-Cmdlet.cs
@@ -45,6 +45,10 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     public partial class ImportEC2SnapshotCmdlet : AmazonEC2ClientCmdlet, IExecutor
     {
         
+        protected override bool IsSensitiveRequest { get; set; } = true;
+        
+        protected override bool IsSensitiveResponse { get; set; } = true;
+        
         #region Parameter ClientData_Comment
         /// <summary>
         /// <para>