+++ modules/AWSPowerShell/Cmdlets/SecurityToken/Basic/Get-STSFederationToken-Cmdlet.cs
@@ -125,6 +125,8 @@ namespace Amazon.PowerShell.Cmdlets.STS
     public partial class GetSTSFederationTokenCmdlet : AmazonSecurityTokenServiceClientCmdlet, IExecutor
     {
         
+        protected override bool IsSensitiveResponse { get; set; } = true;
+        
         #region Parameter DurationInSeconds
         /// <summary>
         /// <para>